raw conversion of CallsBaseMethodsFake specs
